Looooooooooots of debug stuff.  Using , and . to cycle through recorded movement that I store using v
Draw rays for the recorded stuff and spheres for the movement radius
start and end and result.

Align all the little temporary mat4s in use on the stack

Comment out all the sound stuff, something odd is going on with that
stuff.

flymode working a bit better

When not recording, print the footing onscreen

Dumb prints for loaded values not matching previous endpoint.  That
should only happen during a no-collision scenario.  I was half asleep
when I wrote that I think.