I've been using this to test my new (and probably unwise) gpu particles.
Updated todo, not with actual items complete, but with a list of cool
games I spotted while watching the GDC awards shows

I've been working on shaders alot so just using SM5 in the compiles.
Also removed hard coded path in the working dir.

I was never properly setting a layout for the sky draw.  Somehow it was
working sometimes.

Grabbing pointers to the render states I use and set em when needed for
particles / meshes / 2d etc

Make a test particle emitter with alot of stuff swirling about the
origin.